+++ src/components/radio/radio.tsx
@@ -1,4 +1,15 @@
-import { Component, ComponentInterface, h, Host, Prop, Element, Event, EventEmitter, Listen } from '@stencil/core';
+import {
+    Component,
+    ComponentInterface,
+    h,
+    Host,
+    Prop,
+    Element,
+    Event,
+    EventEmitter,
+    Listen,
+    Watch
+} from '@stencil/core';
 import { RadioChosedEvent } from './radio-interface';
 import { RadioGroupMode } from '../radio-group/radio-group-interface';
 import { isEnterKey, isSpaceKey } from '../../utils/helpers';
@@ -6,7 +17,9 @@ import { isEnterKey, isSpaceKey } from '../../utils/helpers';
 @Component({
     tag: 'wcs-radio',
     styleUrl: 'radio.scss',
-    shadow: true
+    shadow: {
+        delegatesFocus: true
+    }
 })
 export class Radio implements ComponentInterface {
     private inputId = `wcs-rb-${radioButtonIds++}`;
@@ -17,7 +30,7 @@ export class Radio implements ComponentInterface {
     @Prop({ mutable: true, reflect: true }) value: any | any[] | undefined | null;
     @Prop({ mutable: true, reflect: true }) label: string;
     /**
-     * If `true`, the radio is selected.
+     * If `true`, the radio is selected. 
      */
     @Prop({mutable: true, reflect: true}) checked = false;
     /**
@@ -26,12 +39,23 @@ export class Radio implements ComponentInterface {
     @Prop({ mutable: true }) disabled = false;
     // FIXME renommer l'évènement c'est pas un onclick mais un onchange
     @Event({eventName: 'wcsRadioClick' }) wcsRadioClick: EventEmitter<RadioChosedEvent>
+    
+    @Prop() name: string;
 
     @Listen('keydown')
     onKeyDown(_event: KeyboardEvent) {
         if ((isSpaceKey(_event) || isEnterKey(_event)) && !this.el.checked) {
-            this.el.checked = true;
-            this.emitRadioChangeEvent();
+            this.checked = true;
+            this.inputEl.click(); // input[radio].checked = true does not trigger any event => input[radio].click() emit a change event
+        }
+    }
+    
+    @Watch("checked")
+    checkedChanged(newValue: boolean) {
+        if(newValue) {
+            this.inputEl.click();
+        } else {
+            this.inputEl.click();
         }
     }
 
@@ -59,16 +83,19 @@ export class Radio implements ComponentInterface {
     }
 
     render() {
+        
         return (
-            <Host slot="option">
+            <Host slot="option" {...(this.disabled ? {tabIndex: -1} : {})}>
                 <input
                     id={this.inputId}
                     type="radio"
+                    name={this.name}
                     value={this.value}
-                    checked={this.checked}
+                    checked={this.checked} // Initial checked state of native input
                     disabled={this.disabled}
                     aria-disabled={this.disabled ? 'true' : null}
-                    aria-checked={`${this.checked}`}/>
+                    aria-checked={`${this.checked}`}
+                />
                 <label htmlFor={`${this.inputId}`}>{this.label}</label>
             </Host>
         );